Initial commit
This is the initial commit of the googleapiclient superdependency. This
wraps all of the dependencies that are required to access google API's
from Python into a single Sublime dependency.

The load order of the dependency is the highest possible so that its
versions of things will be found after other dependencies that are more
targeted, so as to not step on the toes of other libraries.